refactor: Codemie page (#548)
