+++ src/pages/configuration/pages/codemie/view.tsx
@@ -1,66 +1,26 @@
-import { Icon } from '@iconify/react';
 import { EditorDialog } from '@kinvolk/headlamp-plugin/lib/CommonComponents';
 import { KubeObjectInterface } from '@kinvolk/headlamp-plugin/lib/lib/k8s/cluster';
-import {
-  Accordion,
-  AccordionDetails,
-  AccordionSummary,
-  Button,
-  Grid,
-  Paper,
-  Stack,
-  Tooltip,
-  Typography,
-  useTheme,
-} from '@mui/material';
+import { Grid } from '@mui/material';
 import React from 'react';
-import { EmptyList } from '../../../../components/EmptyList';
-import { ErrorContent } from '../../../../components/ErrorContent';
-import { LoadingWrapper } from '../../../../components/LoadingWrapper';
-import { StatusIcon } from '../../../../components/StatusIcon';
 import { CRUD_TYPES } from '../../../../constants/crudTypes';
 import { useResourceCRUDMutation } from '../../../../hooks/useResourceCRUDMutation';
-import { ICONS } from '../../../../icons/iconify-icons-mapping';
-import { CodemieKubeObject } from '../../../../k8s/groups/EDP/Codemie';
 import { CodemieApplicationKubeObject } from '../../../../k8s/groups/EDP/CodemieApplication';
 import { CodemieApplicationKubeObjectConfig } from '../../../../k8s/groups/EDP/CodemieApplication/config';
 import { CodemieApplicationKubeObjectInterface } from '../../../../k8s/groups/EDP/CodemieApplication/types';
 import { CodemieProjectSettingsKubeObject } from '../../../../k8s/groups/EDP/CodemieProjectSettings';
 import { CodemieProjectSettingsKubeObjectConfig } from '../../../../k8s/groups/EDP/CodemieProjectSettings/config';
 import { CodemieProjectSettingsKubeObjectInterface } from '../../../../k8s/groups/EDP/CodemieProjectSettings/types';
-import { getForbiddenError } from '../../../../utils/getForbiddenError';
-import { rem } from '../../../../utils/styling/rem';
 import { ManageCodeMie } from '../../../../widgets/ManageCodeMie';
 import { ConfigurationPageContent } from '../../components/ConfigurationPageContent';
+import { CodemieApplications } from './components/Applications';
+import { CodemieSection } from './components/Codemie';
+import { CodemieProjectSettings } from './components/CodemieProjectSettings';
 import { pageDescription } from './constants';
 import { useTypedPermissions } from './hooks/useTypedPermissions';
 import { useDynamicDataContext } from './providers/DynamicData/hooks';
 
 export const PageView = () => {
-  const {
-    codemie,
-    codemieProject,
-    codemieProjectSettings,
-    codemieQuickLink,
-    codemieSecret,
-    codemieApplications,
-  } = useDynamicDataContext();
-
-  const error =
-    codemie?.error ||
-    codemieProject?.error ||
-    codemieProjectSettings?.error ||
-    codemieQuickLink?.error ||
-    codemieSecret?.error ||
-    codemieApplications?.error;
-
-  const isLoading =
-    codemie?.isLoading ||
-    codemieProject?.isLoading ||
-    codemieProjectSettings?.isLoading ||
-    codemieQuickLink?.isLoading ||
-    codemieSecret?.isLoading ||
-    codemieApplications?.isLoading;
+  const { codemie, codemieQuickLink, codemieSecret } = useDynamicDataContext();
 
   const [isCreateDialogOpen, setCreateDialogOpen] = React.useState<boolean>(false);
 
@@ -68,114 +28,6 @@ export const PageView = () => {
   const handleCloseCreateDialog = () => setCreateDialogOpen(false);
 
   const permissions = useTypedPermissions();
-
-  const renderPageContent = React.useCallback(() => {
-    const forbiddenError = getForbiddenError(error);
-
-    if (forbiddenError) {
-      return <ErrorContent error={forbiddenError} outlined />;
-    }
-
-    if (codemie.error) {
-      return <ErrorContent error={codemie.error} outlined />;
-    }
-
-    if (codemieSecret.error) {
-      return <ErrorContent error={codemieSecret.error} outlined />;
-    }
-
-    if (codemieProject.error) {
-      return <ErrorContent error={codemieProject.error} outlined />;
-    }
-
-    if (!codemie.isLoading && !codemie.data && !codemieSecret.isLoading && !codemieSecret.data) {
-      return (
-        <>
-          <EmptyList
-            customText={'No CodeMie integration found.'}
-            linkText={'Click here to add integration.'}
-            handleClick={handleOpenCreateDialog}
-          />
-        </>
-      );
-    }
-
-    const ownerReference = codemieSecret.data?.metadata?.ownerReferences?.[0]?.kind;
-
-    const status = codemieProject.data?.status?.value;
-    const statusError = codemieProject.data?.status?.error;
-
-    const [icon, color] = CodemieKubeObject.getStatusIcon(status);
-
-    return (
-      <LoadingWrapper isLoading={isLoading}>
-        <Accordion expanded>
-          <AccordionSummary style={{ cursor: 'default' }}>
-            <Typography variant={'h6'}>
-              <Grid container spacing={1} alignItems={'center'}>
-                <Grid item style={{ marginRight: rem(5) }}>
-                  <StatusIcon
-                    icon={icon}
-                    color={color}
-                    Title={
-                      <>
-                        <Typography variant={'subtitle2'} style={{ fontWeight: 600 }}>
-                          {`Status: ${status || 'Unknown'}`}
-                        </Typography>
-                        {!!statusError && (
-                          <Typography variant={'subtitle2'} style={{ marginTop: rem(10) }}>
-                            {statusError}
-                          </Typography>
-                        )}
-                      </>
-                    }
-                  />
-                </Grid>
-                <Grid item>{codemieProject.data?.metadata.name}</Grid>
-                {!!ownerReference && (
-                  <Grid item>
-                    <Tooltip title={`Managed by ${ownerReference}`}>
-                      <Icon
-                        icon={ICONS.CLOUD_LOCK}
-                        width={20}
-                        style={{
-                          display: 'block',
-                        }}
-                      />
-                    </Tooltip>
-                  </Grid>
-                )}
-              </Grid>
-            </Typography>
-          </AccordionSummary>
-          <AccordionDetails>
-            <ManageCodeMie
-              quickLink={codemieQuickLink.data}
-              codemie={codemie.data}
-              codemieSecret={codemieSecret.data}
-              permissions={permissions}
-              handleClosePanel={handleCloseCreateDialog}
-            />
-          </AccordionDetails>
-        </Accordion>
-      </LoadingWrapper>
-    );
-  }, [
-    error,
-    codemie.error,
-    codemie.isLoading,
-    codemie.data,
-    codemieSecret.error,
-    codemieSecret.isLoading,
-    codemieSecret.data,
-    codemieProject.error,
-    codemieProject.data?.status?.value,
-    codemieProject.data?.status?.error,
-    codemieProject.data?.metadata.name,
-    isLoading,
-    codemieQuickLink.data,
-    permissions,
-  ]);
 
   const [editor, setEditor] = React.useState<{
     open: boolean;
@@ -248,8 +100,6 @@ export const PageView = () => {
     }
   };
 
-  const theme = useTheme();
-
   return (
     <ConfigurationPageContent
       creationForm={{
@@ -276,157 +126,16 @@ export const PageView = () => {
     >
       <Grid container spacing={4}>
         <Grid item xs={12}>
-          {renderPageContent()}
+          <CodemieSection
+            handleOpenCreateDialog={handleOpenCreateDialog}
+            handleCloseCreateDialog={handleCloseCreateDialog}
+          />
         </Grid>
         <Grid item xs={12}>
-          <Typography
-            fontSize={theme.typography.pxToRem(24)}
-            color="primary.dark"
-            sx={{ mb: (t) => t.typography.pxToRem(24) }}
-          >
-            Project Settings
-          </Typography>
-          <LoadingWrapper
-            isLoading={codemieProjectSettings.isLoading && !codemieProjectSettings.error}
-          >
-            {codemieProjectSettings.error ? (
-              <ErrorContent error={codemieProjectSettings.error} outlined />
-            ) : codemieProjectSettings.data?.length ? (
-              <Grid container spacing={2}>
-                {codemieProjectSettings.data?.map((_setting) => {
-                  const setting = _setting.jsonData;
-                  const status = setting?.status?.value;
-                  const statusError = setting?.status?.error;
-
-                  const [icon, color] = CodemieProjectSettingsKubeObject.getStatusIcon(status);
-
-                  return (
-                    <Grid item xs={12} key={setting.metadata.name}>
-                      <Paper
-                        sx={{ p: (t) => `${t.typography.pxToRem(10)} ${t.typography.pxToRem(20)}` }}
-                      >
-                        <Stack
-                          direction="row"
-                          spacing={2}
-                          alignItems="center"
-                          justifyContent="space-between"
-                        >
-                          <Stack direction="row" spacing={2} alignItems="center">
-                            <StatusIcon
-                              icon={icon}
-                              color={color}
-                              Title={
-                                <>
-                                  <Typography variant={'subtitle2'} style={{ fontWeight: 600 }}>
-                                    {`Status: ${status || 'Unknown'}`}
-                                  </Typography>
-                                  {!!statusError && (
-                                    <Typography
-                                      variant={'subtitle2'}
-                                      style={{ marginTop: rem(10) }}
-                                    >
-                                      {statusError}
-                                    </Typography>
-                                  )}
-                                </>
-                              }
-                            />
-                            <Typography variant={'h6'}>{setting.metadata.name}</Typography>
-                          </Stack>
-                          <Button
-                            startIcon={<Icon icon={ICONS.PENCIL} />}
-                            size="small"
-                            component={'button'}
-                            style={{ flexShrink: 0 }}
-                            color="inherit"
-                            onClick={() => handleOpenEditor(setting)}
-                          >
-                            Edit YAML
-                          </Button>
-                        </Stack>
-                      </Paper>
-                    </Grid>
-                  );
-                })}
-              </Grid>
-            ) : (
-              <EmptyList customText={'No CodeMie Project Settings Found.'} />
-            )}
-          </LoadingWrapper>
+          <CodemieProjectSettings handleOpenEditor={handleOpenEditor} />
         </Grid>
         <Grid item xs={12} sx={{ pb: (t) => t.typography.pxToRem(40) }}>
-          <Typography
-            fontSize={theme.typography.pxToRem(24)}
-            color="primary.dark"
-            sx={{ mb: (t) => t.typography.pxToRem(24) }}
-          >
-            Applications
-          </Typography>
-          <LoadingWrapper isLoading={codemieApplications.isLoading && !codemieApplications.error}>
-            {codemieApplications.error ? (
-              <ErrorContent error={codemieApplications.error} outlined />
-            ) : codemieApplications.data?.length ? (
-              <Grid container spacing={2}>
-                {codemieApplications.data?.map((_application) => {
-                  const application = _application.jsonData;
-                  const status = application?.status?.value;
-                  const statusError = application?.status?.error;
-
-                  const [icon, color] = CodemieApplicationKubeObject.getStatusIcon(status);
-
-                  return (
-                    <Grid item xs={12} key={application.metadata.name}>
-                      <Paper
-                        sx={{ p: (t) => `${t.typography.pxToRem(10)} ${t.typography.pxToRem(20)}` }}
-                      >
-                        <Stack
-                          direction="row"
-                          spacing={2}
-                          alignItems="center"
-                          justifyContent="space-between"
-                        >
-                          <Stack direction="row" spacing={2} alignItems="center">
-                            <StatusIcon
-                              icon={icon}
-                              color={color}
-                              Title={
-                                <>
-                                  <Typography variant={'subtitle2'} style={{ fontWeight: 600 }}>
-                                    {`Status: ${status || 'Unknown'}`}
-                                  </Typography>
-                                  {!!statusError && (
-                                    <Typography
-                                      variant={'subtitle2'}
-                                      style={{ marginTop: rem(10) }}
-                                    >
-                                      {statusError}
-                                    </Typography>
-                                  )}
-                                </>
-                              }
-                            />
-                            <Typography variant={'h6'}>{application.metadata.name}</Typography>
-                          </Stack>
-                          <Button
-                            startIcon={<Icon icon={ICONS.PENCIL} />}
-                            size="small"
-                            component={'button'}
-                            style={{ flexShrink: 0 }}
-                            color="inherit"
-                            onClick={() => handleOpenEditor(application)}
-                          >
-                            Edit YAML
-                          </Button>
-                        </Stack>
-                      </Paper>
-                    </Grid>
-                  );
-                })}
-              </Grid>
-            ) : (
-              <EmptyList customText={'No CodeMie Applications Found.'} />
-            )}
-          </LoadingWrapper>
+          <CodemieApplications handleOpenEditor={handleOpenEditor} />
         </Grid>
       </Grid>
       {editor.open && editor.data && (
